+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* Consolidated the controller's two "permitted params" methods into one
+
+
 ## [0.0.5] - 2020-06-01
 
 ### Changed